requirements.txt
# Core dependencies
streamlit>=1.28.0
assemblyai>=0.29.0
deepl>=1.17.0
python-dotenv>=1.0.0
httpx>=0.24.0

# Video processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Audio processing
pydub>=0.25.1

# File handling and utilities
requests>=2.31.0
tqdm>=4.66.0
pathlib>=1.0.1
tempfile

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Web interface enhancements
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5
streamlit-player>=0.1.5

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Logging and monitoring
loguru>=0.7.0

# Type hints
typing-extensions>=4.7.0

# Security
cryptography>=41.0.0